requirements/production.txt
Django==2.1.2
uwsgi==2.0.15
mysqlclient==1.3.12
django-redis==4.8.0
django-redis-sessions==0.6.1
requests==2.18.4
jsonfield==2.0.2
web3==4.7.2
simpleflake==0.1.5
python-dateutil==2.1
djangorestframework==3.8.2
drf-yasg==1.10.2
celery==4.1.0
lxml==4.2.1
git+https://github.com/redis/hiredis-py.git@2636dbaf32377ac33543f037abe49590e0d2bc5b
